--- .gradle/2.10/taskArtifacts/cache.properties
@@ -1,1 +1,0 @@
-#Sun Jul 10 22:26:26 EDT 2016
