--- .idea/libraries/mysql_connector_java_5_0_8_bin.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="mysql-connector-java-5.0.8-bin">
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/app/libs/mysql-connector-java-5.0.8-bin.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
