--- .idea/libraries/support_annotations_22_1_1.xml
@@ -1,11 +1,0 @@
-<component name="libraryTable">
-  <library name="support-annotations-22.1.1">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/AppData/Local/Android/Sdk/extras/android/m2repository/com/android/support/support-annotations/22.1.1/support-annotations-22.1.1.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/AppData/Local/Android/Sdk/extras/android/m2repository/com/android/support/support-annotations/22.1.1/support-annotations-22.1.1-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
