+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.ale.logic_package" >
+    package="com.example.ale.applicaiton_layer_package" >
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -11,7 +11,7 @@
         android:theme="@style/Theme.AppCompat.Light.NoActionBar" >
 
         <activity
-            android:name="com.example.ale.logic_package.appController"
+            android:name=".appController"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
