+++ app/src/main/java/client_layer_package/Tour.java
@@ -1,0 +1,7 @@
+package client_layer_package;
+
+/**
+ * Created by Ale on 6/30/2016.
+ */
+public class Tour {
+}
