+++ app/src/main/java/client_layer_package/User.java
@@ -1,0 +1,12 @@
+package client_layer_package;
+
+/**
+ * Created by Ale on 6/30/2016.
+ */
+public class User {
+
+
+
+
+
+}
