--- app/src/main/java/com/example/ale/logic_package/User.java
@@ -1,12 +1,0 @@
-package com.example.ale.logic_package;
-
-/**
- * Created by Ale on 6/30/2016.
- */
-public class User {
-
-
-
-
-
-}
