--- app/src/main/java/com/example/ale/logic_package/appController.java
@@ -1,31 +1,0 @@
-package com.example.ale.logic_package;
-
-import android.app.Activity;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-import android.view.Window;
-import android.widget.EditText;
-
-import storage_package.databaseConnection;
-
-import java.util.concurrent.ExecutionException;
-
-
-public class appController extends Activity {
-
-
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        this.requestWindowFeature(Window.FEATURE_NO_TITLE);
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_log_in2);
-
-    }
-
-
-
-
-}
