+++ app/src/main/java/database_layer_package/DatabaseRead.java
@@ -1,0 +1,7 @@
+package database_layer_package;
+
+/**
+ * Created by Ale on 6/30/2016.
+ */
+public class DatabaseRead {
+}
