--- app/src/main/java/interface_package/BrowseTours.java
@@ -1,7 +1,0 @@
-package interface_package;
-
-/**
- * Created by Ale on 7/10/2016.
- */
-public class BrowseTours {
-}
