--- app/src/main/java/interface_package/LandingPage.java
@@ -1,7 +1,0 @@
-package interface_package;
-
-/**
- * Created by Ale on 7/10/2016.
- */
-public class LandingPage {
-}
