--- app/src/main/java/storage_package/DatabaseRead.java
@@ -1,7 +1,0 @@
-package storage_package;
-
-/**
- * Created by Ale on 6/30/2016.
- */
-public class DatabaseRead {
-}
