--- app/src/main/java/storage_package/DatabaseUpdate.java
@@ -1,7 +1,0 @@
-package storage_package;
-
-/**
- * Created by Ale on 6/30/2016.
- */
-public class DatabaseUpdate {
-}
